+++ programs/winedbg/be_i386.c
@@ -496,7 +496,7 @@ static int be_i386_adjust_pc_for_break(CONTEXT* ctx, BOOL way)
 }
 
 static int be_i386_fetch_integer(const struct dbg_lvalue* lvalue, unsigned size,
-                                 unsigned ext_sign, long long int* ret)
+                                 unsigned ext_sign, LONGLONG* ret)
 {
     if (size != 1 && size != 2 && size != 4 && size != 8) return FALSE;
 
@@ -509,7 +509,7 @@ static int be_i386_fetch_integer(const struct dbg_lvalue* lvalue, unsigned size,
     /* propagate sign information */
     if (ext_sign && size < 8 && (*ret >> (size * 8 - 1)) != 0)
     {
-        long long unsigned int neg = -1;
+        ULONGLONG neg = -1;
         *ret |= neg << (size * 8);
     }
     return TRUE;
